Added building of individual chapters

+++ Build/build_workbook.py
@@ -15,7 +15,7 @@ def path_for_chapter_locale_list(mod, chapter, locale_list):
     # FIXME: When we start localizing, this will need to be smarter
     return os.path.join(mod_dir, mod, '{}-en_US.tex'.format(chapter))
 
-def build_book(book_id, locale_list, paper_size):
+def build_book(book_id, locale_list, paper_size, draft):
     output_tex_path = 'workbook-{}-{}.tex'.format(book_id, locale_list[0])
     output_pdf_path = 'workbook-{}-{}.pdf'.format(book_id, locale_list[0])
     final_pdf_path = '../{}'.format(output_pdf_path)
@@ -63,18 +63,18 @@ def build_book(book_id, locale_list, paper_size):
     output_tex.close()
     os.system('lualatex {}'.format(output_tex_path))
 
-    # If a pdf was made, run it again to get cross-references right
-    if os.path.exists(output_pdf_path):
-        os.system('lualatex {}'.format(output_tex_path))
-        shutil.move(output_pdf_path, final_pdf_path)
-    else:
-        print('Build failed')
-    
-                        
+    if not draft:
+        # If a pdf was made, run it again to get cross-references right
+        if not draft and os.path.exists(output_pdf_path):
+            os.system('lualatex {}'.format(output_tex_path))
+            shutil.move(output_pdf_path, final_pdf_path)
+        else:
+            print('Build failed')
+                
 if not os.path.exists('Intermediate'):
     os.mkdir('Intermediate')
 
-if not os.path.exists('build.cfg'):
+if not os.path.exists('user.cfg'):
     shutil.copyfile('Support/default.cfg', 'user.cfg')
 
 if len(sys.argv) < 2:
@@ -94,5 +94,5 @@ print ('Building workbook ', book)
 
 os.chdir('Intermediate')
 
-build_book(book, ['en_US', 'it_IT'], 'Letter')
+build_book(book, ['en_US', 'it_IT'], 'Letter', false)
